Moved dashlaneCsv types to own file
